deprecation fix

+++ src/Zend/Locale/Math/PhpMath.php
@@ -191,8 +191,8 @@ class Zend_Locale_Math_PhpMath extends Zend_Locale_Math
         if ($op2 == 0) {
             return null;
         }
-        $result = $op1 % $op2;
-        if (is_nan($result) or (($op1 - $result) % $op2 != 0)) {
+        $result = (int) $op1 % $op2;
+        if (is_nan($result) or ((int) ($op1 - $result) % $op2 != 0)) {
             throw new Zend_Locale_Math_Exception("modulus calculation error: $op1 % $op2 != $result", $op1, $op2, $result);
         }
